requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.10
numpy>=1.20.0
pynput>=1.7.0
screeninfo>=0.8
sounddevice>=0.4.0
scipy>=1.7.0
openai-whisper>=20231117
rapidfuzz>=3.0.0
soundfile>=0.12.0